refactor(imports): move multi-sheet importer to settings, drop /imports section
- Add 'Importar datos' settings section hosting the multi-sheet importer (ADMIN)
- Remove the standalone /imports page and its sidebar/route-map entry
- Remove the product-only ImportSection from inventory (superseded by multi-hoja)
- Delete dead ImportSection component
- Update Sidebar/sections/layout tests; frontend 306/307 (1 pre-existing)
