+++ frontend/src/app/settings/sections.test.ts
@@ -2,11 +2,12 @@ import { describe, expect, it } from "vitest";
 import { settingsSections } from "./sections";
 
 describe("settings sections registry", () => {
-  it("defines the three approved settings sections in the required order", () => {
+  it("defines the approved settings sections in the required order", () => {
     expect(settingsSections.map((section) => section.key)).toEqual([
       "general",
       "invoicing",
       "team",
+      "data",
     ]);
   });
